refactor: centralize Sentry context building
