+++ packages/shared/src/triggers/sentry/context.ts
@@ -2,14 +2,17 @@
  * Build context blocks for Sentry automation events.
  */
 
+import type {
+  SentryIssueAlertPayload,
+  SentryIssueWebhookPayload,
+  SentryMetricAlertPayload,
+} from "./payloads";
+
 const MAX_STACK_FRAMES = 5;
 
-export function buildSentryContextBlock(payload: Record<string, unknown>): string {
-  const data = payload.data as Record<string, unknown>;
-  const event = data.event as Record<string, unknown>;
-  const issue = data.issue as Record<string, unknown>;
-  const project = issue.project as Record<string, unknown>;
-  const metadata = event.metadata as Record<string, unknown>;
+export function buildSentryContextBlock(payload: SentryIssueAlertPayload): string {
+  const { event, issue } = payload.data;
+  const { metadata } = event;
 
   const title =
     metadata.type && metadata.value ? `${metadata.type}: ${metadata.value}` : String(issue.title);
@@ -18,7 +21,7 @@ export function buildSentryContextBlock(payload: Record<string, unknown>): strin
     "This automation was triggered by a new Sentry error.",
     "",
     `Error: ${title}`,
-    `Project: ${project.slug}`,
+    `Project: ${issue.project.slug}`,
     `Level: ${issue.level}`,
     `Issue: ${issue.shortId}`,
     `First seen: ${issue.firstSeen}`,
@@ -27,14 +30,9 @@ export function buildSentryContextBlock(payload: Record<string, unknown>): strin
   ];
 
   // Stack trace
-  const exception = event.exception as { values?: Array<Record<string, unknown>> } | undefined;
+  const exception = event.exception;
   if (exception?.values?.length) {
-    const lastException = exception.values[exception.values.length - 1];
-    const stacktrace = lastException.stacktrace as
-      | {
-          frames?: Array<Record<string, unknown>>;
-        }
-      | undefined;
+    const stacktrace = exception.values.at(-1)?.stacktrace;
 
     if (stacktrace?.frames?.length) {
       // Frames are bottom-to-top in Sentry; reverse for most recent first
@@ -53,11 +51,45 @@ export function buildSentryContextBlock(payload: Record<string, unknown>): strin
   }
 
   // Tags
-  const tags = event.tags as Array<{ key: string; value: string }> | undefined;
+  const tags = event.tags;
   if (tags?.length) {
     lines.push("");
     lines.push(`Tags: ${tags.map((t) => `${t.key}=${t.value}`).join(", ")}`);
   }
 
+  return lines.join("\n");
+}
+
+export function buildSentryIssueWebhookContextBlock(payload: SentryIssueWebhookPayload): string {
+  const issue = payload.data.issue;
+  const lines = [
+    "This automation was triggered by a new Sentry issue.",
+    "",
+    `Error: ${issue.title}`,
+    `Project: ${issue.project.slug}`,
+    `Level: ${issue.level}`,
+    `Issue: ${issue.shortId}`,
+  ];
+
+  if (issue.firstSeen) lines.push(`First seen: ${issue.firstSeen}`);
+  if (issue.count !== undefined) lines.push(`Events: ${issue.count}`);
+  if (issue.culprit) lines.push(`Culprit: ${issue.culprit}`);
+  if (issue.web_url) lines.push(`URL: ${issue.web_url}`);
+
+  return lines.join("\n");
+}
+
+export function buildSentryMetricContextBlock(payload: SentryMetricAlertPayload): string {
+  const alert = payload.data.metric_alert;
+  const lines = [
+    "This automation was triggered by a Sentry metric alert.",
+    "",
+    `Alert: ${alert.title}`,
+    `Trigger: ${alert.current_trigger.label}`,
+    `Started: ${alert.date_started}`,
+    `URL: ${payload.data.web_url}`,
+    "",
+    `Description: ${payload.data.description_text}`,
+  ];
   return lines.join("\n");
 }
